requirements.txt
boto3==1.22.1
pyyaml
s3fs
pandas
requests
streamlit
